all get cmds get their own file to
